Add algorithm to save data in cache
Add CharacterCacheDataSource to add function for the LocalDataSource
and save the data received in the database

Change the character id from Int to Long, because the database use the
primaryKey as Long

Remove CharacterLocalDataSourceQualifier and
CharacterRemoteDataSourceQualifier as annotation class, because the
LocalDataSource now has a interface for cache
(CharacterCacheDataSource)

Add function in the Dao to save the Character and SearchPage

Add in the ExceptionExt to the answerBy utils function, a new HOF to
call and create a cache with the response receive from the remote

+++ cache/src/main/java/academy/mukandrew/randm/cache/characters/daos/CharacterDao.kt
@@ -8,7 +8,7 @@ import androidx.room.Query
 @Dao
 internal interface CharacterDao : BaseDao<CharacterEntity> {
     @Query(value = "SELECT * FROM CharacterEntity WHERE id IN (:ids)")
-    suspend fun getCharactersById(ids: List<Int>): List<CharacterEntity>
+    suspend fun getCharactersById(ids: List<Long>): List<CharacterEntity>
 
     @Query(value = "SELECT * FROM CharacterEntity WHERE id = :id")
     suspend fun getCharacterById(id: Int): List<CharacterEntity>
