+++ cache/src/main/java/academy/mukandrew/randm/cache/characters/entities/CharacterSearchPageWithPageNumber.kt
@@ -1,7 +1,7 @@
 package academy.mukandrew.randm.cache.characters.entities
 
 internal data class CharacterSearchPageWithPageNumber(
-    val searchPageId: Int,
+    val searchPageId: Long,
     val name: String,
     val status: String,
     val species: String,
@@ -12,5 +12,5 @@ internal data class CharacterSearchPageWithPageNumber(
     val pages: Int,
     val next: Int,
     val prev: Int,
-    val characterIds: List<Int>
+    val characterIds: List<Long>
 )
