+++ domain/src/main/java/academy/mukandrew/randm/domain/characters/usecases/abstractions/GetMultipleCharacterUseCase.kt
@@ -4,5 +4,5 @@ import academy.mukandrew.randm.common.result.Answer
 import academy.mukandrew.randm.domain.characters.models.Character
 
 abstract class GetMultipleCharacterUseCase {
-    abstract suspend operator fun invoke(ids: List<Int>): Answer<List<Character>>
+    abstract suspend operator fun invoke(ids: List<Long>): Answer<List<Character>>
 }
