+++ remote/src/main/java/academy/mukandrew/randm/remote/characters/mappers/queries/CharactersByIdsQueryMapper.kt
@@ -2,6 +2,6 @@ package academy.mukandrew.randm.remote.characters.mappers.queries
 
 import academy.mukandrew.randm.remote.CharactersByIdsQuery
 
-fun getCharactersByIdsQuery(ids: List<Int>): CharactersByIdsQuery {
+fun getCharactersByIdsQuery(ids: List<Long>): CharactersByIdsQuery {
     return CharactersByIdsQuery(ids.map { it.toString() })
 }
